fix: segmentation violation patched by disabling bdwgc
This is obviously not ideal and we should investigate more. GC_DONT_GC=1
didn't seem to fix anything, also, for what it's worth. Seems like the
failure was due to an infinite loop of tree modifications relate to
nix::Settings::Settings() (valgrind and lldb and gdb all seemed to imply
this).